Added range ring annotation
